Module Odoc_model.Comment

module Path = Paths.Path

module Reference = Paths.Reference

module Identifier = Paths.Identifier

type 'a with_location = 'a Location_.with_location

type style = [

| `Bold

| `Italic

| `Emphasis

| `Superscript

| `Subscript

]

type alignment = [

| `Left

| `Center

| `Right

]

type media = [

| `Image

| `Audio

| `Video

]

type raw_markup_target = string

type leaf_inline_element = [

| `Space

| `Word of string

| `Code_span of string

| `Math_span of string

| `Raw_markup of raw_markup_target * string

]

type non_link_inline_element = [

| leaf_inline_element

| `Styled of style * non_link_inline_element with_location list

]

type link_content = non_link_inline_element with_location list

type reference_element = [

| `Reference of Reference.t * link_content

]

type inline_element = [

| leaf_inline_element

| `Styled of style * inline_element with_location list

| reference_element

| `Link of string * link_content

]

type paragraph = inline_element with_location list

type module_reference = {

module_reference : Reference.Module.t;

module_synopsis : paragraph option;

}

The {!modules: ...} markup. module_synopsis is initially None, it is resolved during linking.

type 'a cell = 'a with_location list * [ `Header | `Data ]

type 'a row = 'a cell list

type 'a grid = 'a row list

type 'a abstract_table = {

data : 'a grid;

align : alignment option list option;

}

type media_href = [

| `Link of string

| `Reference of Reference.Asset.t

]

type media_element = [

| `Media of media_href * media * string

]

type nestable_block_element = [

| `Paragraph of paragraph

| `Code_block of string option * string with_location * nestable_block_element with_location list option

| `Math_block of string

| `Verbatim of string

| `Modules of module_reference list

| `Table of nestable_block_element abstract_table

| `List of [ `Unordered | `Ordered ] * nestable_block_element with_location list list

| media_element

]

type tag = [

| `Author of string

| `Deprecated of nestable_block_element with_location list

| `Param of string * nestable_block_element with_location list

| `Raise of [ `Code_span of string | reference_element ] * nestable_block_element with_location list

| `Return of nestable_block_element with_location list

| `See of [ `Url | `File | `Document ] * string * nestable_block_element with_location list

| `Since of string

| `Before of string * nestable_block_element with_location list

| `Version of string

| `Alert of string * string option

]

type heading_level = [

| `Title

| `Section

| `Subsection

| `Subsubsection

| `Paragraph

| `Subparagraph

]

type attached_block_element = [

| nestable_block_element

| `Tag of tag

]

type heading_attrs = {

heading_level : heading_level;

heading_label_explicit : bool;
(*
Whether the label have been written by the user.
*)

}

type block_element = [

| nestable_block_element

| `Heading of heading_attrs * Identifier.Label.t * inline_element with_location list

| `Tag of tag

]

type elements = block_element with_location list

type docs = {

elements : elements;

suppress_warnings : bool;

}

type docs_or_stop = [

| `Docs of docs

| `Stop

]

val synopsis : [> `Paragraph of 'a ] Location_.with_location list -> 'a option

The synopsis is the first element of a comment if it is a paragraph. Otherwise, there is no synopsis.

val link_content_of_inline_element : inline_element with_location -> link_content

val link_content_of_inline_elements : inline_element with_location list -> non_link_inline_element with_location list

val find_zero_heading : [> `Heading of heading_attrs * 'a * inline_element with_location list ] Location_.with_location Odoc_utils.List.t -> link_content option

val to_string : link_content -> string