Module Odoc_model.Location_

include module type of struct include Odoc_parser.Loc end

This module concerns locations in source files, both points indicating a specific character and spans between two points.

Basic types

type point = Odoc_parser.Loc.point = {

line : int;

column : int;

}

A specific character

type span = Odoc_parser.Loc.span = {

file : string;

start : point;

end_ : point;

}

A range of characters between start and end_ in a particular file

val span : span list -> span

span spans takes a list of spans and returns a single span starting at the start of the first span and ending at the end of the final span

val nudge_start : int -> span -> span

This adjusts only the column number, implicitly assuming that the offset does not move the location across a newline character.

val nudge_end : int -> span -> span

This adjusts only the column number, implicitly assuming that the offset does not move the location across a newline character.

Located values

type +'a with_location = 'a Odoc_parser.Loc.with_location = {

location : span;

value : 'a;

}

Describes values located at a particular span

val at : span -> 'a -> 'a with_location

Constructor for with_location

val location : 'a with_location -> span

Returns the location of a located value

val value : 'a with_location -> 'a

Returns the value of a located value

val map : ('a -> 'b) -> 'a with_location -> 'b with_location

Map over a located value without changing its location

val same : _ with_location -> 'b -> 'b with_location

same x y retuns the value y wrapped in a with_location whose location is that of x

val spans_multiple_lines : _ with_location -> bool

spans_multiple_lines x checks to see whether x is located on a single line or whether it covers more than one.

val set_end_as_offset_from_start : int -> span -> span

val in_string : string -> offset:int -> length:int -> span -> span

val pp : Format.formatter -> span -> unit

val pp_span_start : Format.formatter -> span -> unit

val span_equal : span -> span -> bool