Module Identifier.Mk

val page : (ContainerPage.t option * Names.PageName.t) -> [> `Page of ContainerPage.t option * Names.PageName.t ] id

val leaf_page : (ContainerPage.t option * Names.PageName.t) -> [> `LeafPage of ContainerPage.t option * Names.PageName.t ] id

val source_page : (ContainerPage.t * string) -> SourcePage.t

val asset_file : (Page.t * Names.AssetName.t) -> AssetFile.t

val root : (ContainerPage.t option * Names.ModuleName.t) -> [> `Root of ContainerPage.t option * Names.ModuleName.t ] id

val implementation : string -> [> `Implementation of Names.ModuleName.t ] id

val module_ : (Signature.t * Names.ModuleName.t) -> [> `Module of Signature.t * Names.ModuleName.t ] id

val parameter : (Signature.t * Names.ModuleName.t) -> [> `Parameter of Signature.t * Names.ModuleName.t ] id

val result : Signature.t -> [> `Result of Signature.t ] id

val module_type : (Signature.t * Names.ModuleTypeName.t) -> [> `ModuleType of Signature.t * Names.ModuleTypeName.t ] id

val class_ : (Signature.t * Names.TypeName.t) -> [> `Class of Signature.t * Names.TypeName.t ] id

val class_type : (Signature.t * Names.TypeName.t) -> [> `ClassType of Signature.t * Names.TypeName.t ] id

val type_ : (Signature.t * Names.TypeName.t) -> [> `Type of Signature.t * Names.TypeName.t ] id

val core_type : string -> [> `CoreType of Names.TypeName.t ] id

val constructor : (DataType.t * Names.ConstructorName.t) -> [> `Constructor of DataType.t * Names.ConstructorName.t ] id

val field : (FieldParent.t * Names.FieldName.t) -> [> `Field of FieldParent.t * Names.FieldName.t ] id

val extension : (Signature.t * Names.ExtensionName.t) -> [> `Extension of Signature.t * Names.ExtensionName.t ] id

val extension_decl : (Signature.t * (Names.ExtensionName.t * Names.ExtensionName.t)) -> [> `ExtensionDecl of Signature.t * Names.ExtensionName.t * Names.ExtensionName.t ] id

extension_decl (sg, e1, eN) defines an extension declaration where sg is the parent, e1 is the first constructor of the extension, and eN is the constructor the Id is created for. e1 will be used for the url, and eN will be the one displayed. The first constructor of the extension will always be used to reference the extension point.

val exception_ : (Signature.t * Names.ExceptionName.t) -> [> `Exception of Signature.t * Names.ExceptionName.t ] id

val value : (Signature.t * Names.ValueName.t) -> [> `Value of Signature.t * Names.ValueName.t ] id

val method_ : (ClassSignature.t * Names.MethodName.t) -> [> `Method of ClassSignature.t * Names.MethodName.t ] id

val instance_variable : (ClassSignature.t * Names.InstanceVariableName.t) -> [> `InstanceVariable of ClassSignature.t * Names.InstanceVariableName.t ] id

val label : (LabelParent.t * Names.LabelName.t) -> [> `Label of LabelParent.t * Names.LabelName.t ] id

val source_location : (SourcePage.t * Names.DefName.t) -> [> `SourceLocation of SourcePage.t * Names.DefName.t ] id

val source_location_mod : SourcePage.t -> [> `SourceLocationMod of SourcePage.t ] id

val source_location_int : (SourcePage.t * Names.LocalName.t) -> [> `SourceLocationInternal of SourcePage.t * Names.LocalName.t ] id